🔀 refactor: move common methods to super class The changes were made to remove unused methods and imports in the ClipboardKeypadView and MainKeyboardView classes. These changes improve code readability and maintainability by removing unnecessary code clutter and moving common functionality to the super class.
